PAD-2 + corrected Big decimals + corrected wrappers to primitives

+++ src/main/java/bg/enterprise/parent_app/model/dto/DosageLogDto.java
@@ -2,13 +2,13 @@ package bg.enterprise.parent_app.model.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import java.math.BigDecimal;
+
 import java.time.Instant;
 
 @Getter
 @Setter
 public class DosageLogDto extends AuditDto {
-    private BigDecimal dosageGiven;
+    private float dosageGiven;
     private Instant administeredAt;
     private String notes;
 }
